requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
extra-streamlit-components==0.1.71
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kagglehub==0.3.0
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.9.3
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyopencl==2024.2.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytools==2024.1.14
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safeloader==0.1.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
smmap==5.0.1
st-theme==1.2.3
streamlit==1.39.0
streamlit-authenticator==0.4.1
streamlit-authenticator-mongo==0.1.5
streamlit-navigation-bar==3.3.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
text-unidecode==1.3
tf-slim==1.1.0
tf_keras==2.17.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
